+++ ietf/doc/tests.py
@@ -36,7 +36,7 @@ import debug                            # pyflakes:ignore
 from ietf.doc.models import ( Document, DocAlias, DocRelationshipName, RelatedDocument, State,
     DocEvent, BallotPositionDocEvent, LastCallDocEvent, WriteupDocEvent, NewRevisionDocEvent, BallotType,
     EditedAuthorsDocEvent )
-from ietf.doc.factories import ( DocumentFactory, DocEventFactory, CharterFactory, 
+from ietf.doc.factories import ( DocumentFactory, DocEventFactory, CharterFactory,
     ConflictReviewFactory, WgDraftFactory, IndividualDraftFactory, WgRfcFactory, 
     IndividualRfcFactory, StateDocEventFactory, BallotPositionDocEventFactory, 
     BallotDocEventFactory, DocumentAuthorFactory, NewRevisionDocEventFactory,
@@ -44,7 +44,6 @@ from ietf.doc.factories import ( DocumentFactory, DocEventFactory, CharterFactor
 from ietf.doc.forms import NotifyForm
 from ietf.doc.fields import SearchableDocumentsField
 from ietf.doc.utils import create_ballot_if_not_open, uppercase_std_abbreviated_name
-from ietf.doc.views_search import ad_dashboard_group, ad_dashboard_group_type, shorten_group_name # TODO: red flag that we're importing from views in tests. Move these to utils.
 from ietf.group.models import Group, Role
 from ietf.group.factories import GroupFactory, RoleFactory
 from ietf.ipr.factories import HolderIprDisclosureFactory
@@ -60,6 +59,8 @@ from ietf.utils.test_utils import login_testing_unauthorized, unicontent
 from ietf.utils.test_utils import TestCase
 from ietf.utils.text import normalize_text
 from ietf.utils.timezone import date_today, datetime_today, DEADLINE_TZINFO, RPC_TZINFO
+from ietf.doc.utils_search import doc_type, doc_state
+from ietf.doc.utils_search import doc_type_name as get_doc_type_name
 
 
 class SearchTests(TestCase):
@@ -297,7 +298,6 @@ class SearchTests(TestCase):
                     elif doc_type_name == 'charter':
                         doc = CharterFactory(ad=ad, states=[(doc_type_name, state)])
                     elif doc_type_name == 'bofreq':
-                        # Note that the view currently doesn't handle bofreqs
                         doc = BofreqFactory(states=[(doc_type_name, state)], bofreqresponsibledocevent__responsible=[ad])
                     elif doc_type_name == 'conflrev':
                         doc = ConflictReviewFactory(ad=ad, states=State.objects.filter(type_id=doc_type_name, slug=state))
@@ -307,14 +307,18 @@ class SearchTests(TestCase):
                         # Currently unreachable
                         doc = DocumentFactory(type_id=doc_type_name, ad=ad, states=[(doc_type_name, state)])
 
-                    if not slugify(ad_dashboard_group_type(doc)) in ('document', 'none'):
-                        expected[(slugify(ad_dashboard_group_type(doc)), slugify(ad.full_name_as_key()), slugify(shorten_group_name(ad_dashboard_group(doc))))] += 1
+                    if not doc_type(doc) in ('document', 'none'):
+                        expected[(slugify(get_doc_type_name(doc_type(doc))), slugify(ad.full_name_as_key()), doc_state(doc))] += 1
         
         url = urlreverse('ietf.doc.views_search.ad_workload')
         r = self.client.get(url)
         self.assertEqual(r.status_code, 200)
         q = PyQuery(r.content)
+        print(r.content)
+        print(expected)
         for group_type, ad, group in expected:
+            print(group_type, ad, group)
+            print(q(f'#{group_type}-{ad}-{group}').text())
             self.assertEqual(int(q(f'#{group_type}-{ad}-{group}').text()),expected[(group_type, ad, group)])
 
     def test_docs_for_ad(self):
